refactor: every ADP column becomes a dense integer rank
The columns held three incompatible units: FFC and MFL publish an average
overall pick over DIFFERENT draft pools (MFL's look like 7.74), FantasyPros
publishes a consensus rank, and adp_formatted carried FFC's round.pick
rendering ("3.05"). Three units on one row cannot be compared by eye, which is
the only reason to show them side by side.

Every ADP column is now a PositiveIntegerField holding a dense 1..N index, 1 =
first player off the board. No rounding, no truncation - the sort index of the
values that exist. A row now reads "board 10 / FFC 10 / MFL 17 / FPROS 21" and
the disagreement is the point.

rerank_year() in services/adp/ranking.py is the reusable pass, and it runs
AUTOMATICALLY rather than being a step anyone has to remember:
refresh_players_from_ffc calls it unconditionally at the end, because a refresh
that inserts a player mid-board otherwise leaves a duplicate or a hole. It is
idempotent, orders by the current value with name as a deterministic tiebreaker,
leaves NULLs NULL, and writes via queryset update() so it cannot trip
Player.save()'s price flooring. sync_source and apply_source assign ranks
directly as they write.

Migrations are split deliberately: 0089 ranks while the columns are still
numeric, 0090 casts to integer. Postgres refuses to ALTER a column type in the
transaction that just wrote to the table, and the order is not optional - a
plain cast would collapse round.pick 1.01 through 1.10 into ten players all
sitting at 1. The three serializers exposing adp_formatted move to IntegerField;
add_missing_players and the legacy update_player_prices no longer write
round.pick strings into what is now an integer column.

Making the columns comparable immediately surfaced something that changes how
MFL should be used: it ranks 7 QBs inside its top 50 where FFC ranks 1, with a
median shift of QB -28, TE -16, WR +16 against FFC. That is superflex/2QB
contamination - MFL is the platform of choice for those formats - and applying
MFL to a 1QB auction would inflate every QB price. Recorded in the source's
caveat (so it shows on the admin confirm page), the provider docstring, and
AGENTS.md. The source is still a good second opinion at RB/WR.

Tests: AdpRerankTests (dense index, idempotence, mid-board insert, NULLs,
deterministic ties, prices untouched, dry run, year isolation) and
FfcRefreshReranksTests (a refresh leaves every column densely ranked and never
stores the feed's round.pick string). 108 total.

+++ draft/api/views/draft.py
@@ -53,7 +53,8 @@ class PlayerOutputSerializer(BaseSerializer):
     player_id = serializers.CharField()
     name = serializers.CharField()
     position = serializers.CharField()
-    adp_formatted = serializers.DecimalField(max_digits=8, decimal_places=2)
+    # Integer RANK (1 = first off the board), not a decimal round.pick.
+    adp_formatted = serializers.IntegerField()
     projected_price = serializers.DecimalField(max_digits=8, decimal_places=2)
     override_price = serializers.DecimalField(max_digits=8, decimal_places=2)
     nickname = serializers.CharField()
@@ -288,7 +289,7 @@ class TargetTierPlayerOutputSerializer(BaseSerializer):
     name = serializers.CharField(source="player.name")
     position = serializers.CharField(source="player.position")
     target_tier = serializers.IntegerField(source="player.target_tier")
-    adp_formatted = serializers.DecimalField(max_digits=8, decimal_places=2, source="player.adp_formatted")
+    adp_formatted = serializers.IntegerField(source="player.adp_formatted")
     favorite = serializers.BooleanField(allow_null=True, source="player.favorite")
     notes = serializers.CharField(allow_null=True, source="player.notes")
     team = serializers.CharField(allow_null=True, source="player.team.code", default=None)
